GSK-58 Project Migration: correct dtos model
