+++ giskard-server/src/main/java/ai/giskard/service/dto/ml/ModelDTO.java
@@ -17,33 +17,11 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @NoArgsConstructor
-public class ModelDTO {
-    @JsonProperty("id")
-    @NotNull
-    protected Long id;
-    @JsonIgnore
-    protected ProjectDTO project;
-    @JsonProperty("name")
-    protected String fileName;
+public class ModelDTO extends FileDTO {
+
     @JsonProperty("python_version")
     private String pythonVersion;
-    @JsonProperty("creation_date")
-    protected LocalDateTime createdOn;
-    protected String location;
-
-    public Long getSize() throws IOException {
-        Path path = Paths.get(location);
-        if (Files.exists(path)) {
-            return Files.size(path);
-        } else {
-            return 0L;
-        }
-    }
-
-    public Long getProjectId() {
-        return project.getId();
-    }
-
+    
     public ModelDTO(ProjectModel model) {
         this.id = model.getId();
         this.fileName = model.getFileName();
