+++ giskard-server/src/main/java/ai/giskard/service/dto/ml/ProjectPostDTO.java
@@ -13,11 +13,17 @@ public class ProjectPostDTO {
     @Getter
     @NotNull
     private Long id;
+
     @Getter
     @NotNull
     private String name;
+
     @Getter
     @NotNull
     private String key;
+
+    @Getter
+    @NotNull
+    private String description;
 
 }
